+++ changelog.md
@@ -2,9 +2,16 @@
 
 ## Unreleased
 
+### Breaking Changes
+
+- Drop support for Node 8
+
 ### Added
 
 - `CP_PAY` submission event to form schema
+- [`createForm()`](./docs/forms.md#createform) function
+- [`updateForm()`](./docs/forms.md#updateform) function
+- [`deleteForm()`](./docs/forms.md#deleteform) function
 
 ### Removed